Sol: Fix Shaman's Earth and Fire Elemental
If the player targets a creature and summons either the Earth or Fire Elemental
it attacks the target at once. Use an appropriate check to determine if the target
can actually be attacked. As it was before the Elemental attacked the target without
distance check and even through walls (e.g. you could pull "The Prophet Tharon'ja"
from below where Trollgore stands).