Sol: Remove worldserver parameter "IsPreloadedContinentTransport.Enabled"
The implementation was wrong as it loaded all grids for gameobjects type 11,
which means type "Transport" like elevators etc., but the continent transports
are type 15 "Motion Transport". The parameter is also not needed because the
continent transports are set active when created and so will automatically
load all grids after one complete cycle. In order to save memory only the
parameter "IsContinentTransport.Enabled" is of use if continent transport
is not needed at all.