Sol: Spawn flying creatures at their spawn point, not on the ground
Remove a part of the code which relocated flying creatures to the
ground when loaded from the DB with their respawn timer set. This
was a very old functionality introduced 2008 with this CMaNGOS commit:
https://github.com/cmangos/mangos-wotlk/commit/6b4737a4dde6fadc2fcdc0fea8e30bc8e23d87f8

This was presumably taken over by TC and finally AC. Sadly, there's
no information concerning the original intent, perhaps this was
necessary for the core at this time. As it is now this leads to a bug
because the flying creatures do not spawn at their spawn point
and start from the ground if for example following a waypoint path.