Sol: Fix SMART_EVENT_OOC_LOS and SMART_EVENT_IC_LOS
Only max cooldown was used for the initial delay. The author
obviously wanted to disable the initial delay and accidentally
also removed the "break". This leads to the subsequent call of
"RecalcTimer" using "e.event.distance.repeat" which has the
same parameter position as "e.event.los.cooldownMax". In order
to not break existing functionality, keep the initial delay,
but now take also "cooldownMin" into account.