Sol: Fix gcc-10 warnings -Wnarrowing
Also remove cmake policy CMP0074 and fix a buggy implementation introduced with AC commit 3d36bb81a4d298508550118b075fda4eddc06b7c:
The achievement category ID cannot be negative, only the parent category can. If the parent category is -1 this means
that the category is a root category and the algorithm can stop searching for a statistics category.