Sol: Fix game event entry value range for creatures, gameobjects and pools
The field "eventEntry" in table "game_event" allows values between 0 and 255 (unsigned tinyint),
the same field in "game_event_creature", "game_event_gameobject" and "game_event_pool" is signed
and allows only values between -128 and 127. This means that not the full range of events can be
addressed in these tables. As a solution the data type is set to "signed smallint" with an
appropriate constraint which checks that "eventEntry" is >= -255, <= 255 and not 0.